Install numpy with apt package manager
To get BLAS accelleration. This will reduce the time to do a 1000x1000
dot product in numpy from 95 seconds to 30 seconds on an S2L camera. It
should also be more robust to get precompiled packages since pip is
dependent on what wheels are build/uploaded for the current release of
numpy.